+++ packages/authentication/modules/user/getUser/index.ts
@@ -5,10 +5,10 @@ import { IJWTContent } from '@baseapp-frontend/utils/types/jwt'
 import { IUser } from '../../../types/user'
 import { GetUserOptions } from './types'
 
-const getUser = <TUser extends Partial<IUser> & IJWTContent>({
+const getUser = async <TUser extends Partial<IUser> & IJWTContent>({
   cookieName = ACCESS_COOKIE_NAME,
 }: GetUserOptions = {}) => {
-  const token = getToken(cookieName)
+  const token = await getToken(cookieName)
   if (token) {
     try {
       const user = decodeJWT<TUser>(token)